requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-limiter==3.5.0
python-deepl==1.16.1
flask-cors==4.0.0 